combine 2 user services into one
